% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{anaconda.Windowsx64}
\alias{anaconda.Windowsx64}
\title{importFrom xml2 read_html html_nodes
importFrom rvest html_nodes html_attr}
\usage{
\method{anaconda}{Windowsx64}(os)
}
\description{
importFrom xml2 read_html html_nodes
importFrom rvest html_nodes html_attr
}
