% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean remove java or python from multilinguer package}
\usage{
clean(what)
}
\arguments{
\item{what}{java or python. No defualt.}
}
\description{
clean remove java or python from multilinguer package
}
