% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conda_available}
\alias{conda_available}
\title{check if possible to find conda}
\usage{
conda_available()
}
\description{
this function depends on find_conda() from reticulate package.
if find_conda() return NULL, conda_available() return FALSE.
}
