% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{corretto.Windowsx64}
\alias{corretto.Windowsx64}
\title{importFrom xml2 read_html html_nodes
importFrom rvest html_nodes html_attr}
\usage{
\method{corretto}{Windowsx64}(os)
}
\description{
importFrom xml2 read_html html_nodes
importFrom rvest html_nodes html_attr
}
