% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correttor-package.R
\docType{package}
\name{correttor-package}
\alias{correttor}
\alias{correttor-package}
\title{correttor: Install Amazon Corretto}
\description{
Provide install function of Amazon Corretto to use rJava for all os.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/installjava}
  \item Report bugs at \url{https://github.com/mrchypark/installjava/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
