% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_corretto.R
\name{crt_download}
\alias{crt_download}
\title{corretto installer download}
\usage{
crt_download(os, dest)
}
\arguments{
\item{os}{Type of OS}

\item{dest}{where to download}
}
\description{
windows x64 is for java11 x64
windows x86 is for java8 x86
mac is for java11 x64
}
