% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_ssl_error}
\alias{fix_ssl_error}
\title{fix ssl error on conda}
\usage{
fix_ssl_error(loc = conda_root())
}
\arguments{
\item{loc}{location of conda}
}
\description{
Solution of this \href{https://github.com/conda/conda/issues/8273}{issue}
}
