% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_conda}
\alias{has_conda}
\title{has conda}
\usage{
has_conda(force = FALSE, update = TRUE, silent = TRUE,
  quiet = FALSE)
}
\arguments{
\item{force}{force installation. Defualt is FALSE}

\item{update}{defualt is T with conda update.}

\item{silent}{defualt is T.}

\item{quiet}{defualt is F.}
}
\description{
Check if there is a conda, and install if it does not exist.
}
