% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_java}
\alias{has_java}
\title{has java}
\usage{
has_java(force = FALSE)
}
\arguments{
\item{force}{force installation. Defualt is FALSE}
}
\description{
Check if there is a java, and install if it does not exist.
}
