% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{inst.Windowsx64}
\alias{inst.Windowsx64}
\title{importFrom xml2 read_html
importFrom rvest html_nodes html_attr}
\usage{
\method{inst}{Windowsx64}(os)
}
\description{
importFrom xml2 read_html
importFrom rvest html_nodes html_attr
}
