% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_miniconda}
\alias{install_miniconda}
\title{Install Miniconda}
\usage{
install_miniconda(path = "", update = TRUE, force = FALSE)
}
\arguments{
\item{path}{The path in which Miniconda will be installed. Note that the
installer does not support paths containing spaces.}

\item{update}{Boolean; update to the latest version of Miniconda after install?}

\item{force}{Boolean; force re-installation if Miniconda is already installed
at the requested path?}
}
\description{
Install Miniconda
}
\details{
Download the \href{https://docs.conda.io/en/latest/miniconda.html}{Miniconda}
installer, and use it to install Miniconda.
All function and descriptions from \href{https://github.com/rstudio/reticulate/blob/master/R/miniconda.R}{reticulate package}
}
