% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installjava-package.R
\docType{package}
\name{installjava-package}
\alias{installjava}
\alias{installjava-package}
\title{installjava: Install Java(Amazon Corretto)}
\description{
Provide Java(Amazon Corretto) install function for all os.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/installjava}
  \item Report bugs at \url{https://github.com/mrchypark/installjava/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
