% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdk_download.R
\name{jdk_download}
\alias{jdk_download}
\title{corretto installer download}
\usage{
jdk_download(dest = "")
}
\arguments{
\item{dest}{target path and file name.}
}
\description{
windows x64 is for java11 x64
windows x86 is for java8 x86
mac is for java11 x64
linux is for java11 x64
}
