% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdk_list.R
\docType{data}
\name{jdk_list}
\alias{jdk_list}
\title{List support jdk}
\format{
list:
\describe{
\item{os type and arc}{return value of get_os()}
\item{jdk version}{11 is main.}
\item{gui mode or not}{zip is main.}
}
}
\usage{
jdk_list
}
\description{
List support jdk
}
\keyword{datasets}
