% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdk_install.R
\name{jdk_path}
\alias{jdk_path}
\title{jdk path}
\usage{
jdk_path()
}
\description{
The path to the Jdk installation to use. By default, an OS-specific
path is used. If you'd like to instead set your own path, you can set the
\code{MULTILINGUER_JDK_PATH} environment variable.
}
\details{
You need to make sure you have permission when setting the path with \code{MULTILINGUER_JDK_PATH}.
}
\examples{
  jdk_path()

}
\seealso{
Other jdk-installer: 
\code{\link{install_jdk}()},
\code{\link{remove_jdk}()}
}
\concept{jdk-installer}
