% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_node.R
\name{node_download}
\alias{node_download}
\title{nodejs installer download}
\usage{
node_download(os, dest)
}
\arguments{
\item{os}{Type of OS}

\item{dest}{where to download}
}
\description{
nodejs installer download
}
