% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdk_install.R
\name{remove_jdk}
\alias{remove_jdk}
\alias{uninstall_jdk}
\title{Remove Jdk}
\usage{
remove_jdk(path = jdk_path())

uninstall_jdk(path = jdk_path())
}
\arguments{
\item{path}{The path in which jdk is installed.}
}
\description{
Uninstall jdk from this package.
}
\examples{
\dontrun{
  remove_jdk()
  uninstall_jdk()
}
}
\seealso{
Other jdk-installer: 
\code{\link{install_jdk}()},
\code{\link{jdk_path}()}

Other jdk-installer: 
\code{\link{install_jdk}()},
\code{\link{jdk_path}()}
}
\concept{jdk-installer}
