% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rust.R
\name{rust_download}
\alias{rust_download}
\title{rust installer download}
\usage{
rust_download(os, dest)
}
\arguments{
\item{os}{Type of OS}

\item{dest}{where to download}
}
\description{
rust installer download
}
