% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rust.R
\name{rustup_windows}
\alias{rustup_windows}
\alias{rust_uninstall}
\title{Install Rust}
\usage{
rustup_windows(targets)

rust_uninstall()
}
\arguments{
\item{targets}{which compiler targets would you like to install}
}
\description{
Install Rust
}
\examples{
\dontrun{
rust_uninstall()
rustup_windows()
}
}
