% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_conda}
\alias{has_conda}
\title{has conda}
\usage{
has_conda(update = TRUE, silent = T, quiet = F)
}
\arguments{
\item{update}{defualt is T with conda update.}

\item{silent}{defualt is T.}

\item{quiet}{defualt is F.}
}
\description{
Check if there is a conda, and install if it does not exist.
}
