% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_conda}
\alias{install_conda}
\alias{install_miniconda}
\title{Install miniconda}
\usage{
install_conda()

install_miniconda()
}
\description{
Install miniconda
}
\details{
Download the \href{https://docs.conda.io/en/latest/miniconda.html}{Miniconda}
installer, and use it to install Miniconda.
All function and descriptions from \href{https://github.com/rstudio/reticulate/blob/master/R/miniconda.R}{reticulate package}
}
\examples{
\dontrun{
  install_conda()
}
}
