% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_java}
\alias{install_java}
\alias{install_jdk}
\title{install conda}
\usage{
install_java()

install_jdk()
}
\description{
install miniconda
}
\details{
install \code{java}

install \code{corretto} which is one of openjdk(java) distro.
Case of \code{MacOS}, remove all java and reinstall corretto version 11.
}
