% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilinguer-package.R
\docType{package}
\name{multilinguer-package}
\alias{multilinguer}
\alias{multilinguer-package}
\title{multilinguer: Install Amazon Corretto and Anaconda for All OS}
\description{
Provide install function of Amazon Corretto to use rJava and Anaconda to use reticulate for all os.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/multilinguer}
  \item Report bugs at \url{https://github.com/mrchypark/multilinguer/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
