% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilinguer-package.R
\docType{package}
\name{multilinguer-package}
\alias{multilinguer}
\alias{multilinguer-package}
\title{multilinguer: Gentle Language Installer for R User}
\description{
Provides install functions of other languages 
             such as 'java', 'python' for windows and macos.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/multilinguer}
  \item Report bugs at \url{https://github.com/mrchypark/multilinguer/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
